fix: Small code cleanups

+++ src/main/java/com/gildedgames/orbis_api/client/gui/util/GuiFrame.java
@@ -259,7 +259,7 @@ public abstract class GuiFrame extends GuiContainer implements IGuiFrame
 
 			if (gui != null)
 			{
-				g.setWorldAndResolution(this.mc, (int) InputHelper.getScreenWidth(), (int) InputHelper.getScreenHeight());
+				g.setWorldAndResolution(this.mc, InputHelper.getScreenWidth(), InputHelper.getScreenHeight());
 			}
 		}
